Date parser refactoring
